requirements.txt
# LangGraph & LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0

# MCP
mcp>=0.9.0

# API 서버
fastapi>=0.109.0
uvicorn>=0.27.0

# 데이터베이스
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0

# 미들웨어 & 안전장치
guardrails-ai>=0.4.0

# 유틸
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
